Removes concrete impl from Entity
