+++ src/pyrite/entity/__init__.py
@@ -1,0 +1,1 @@
+from .entity import BaseEntity as Entity  # noqa: F401
